Extract models to model parser folder
