+++ swagger_parser/lib/src/generator/models/programming_language.dart
@@ -1,6 +1,8 @@
 import 'package:collection/collection.dart';
 
 import '../../parser/model/open_api_info.dart';
+import '../../parser/model/universal_data_class.dart';
+import '../../parser/model/universal_rest_client.dart';
 import '../generator_exception.dart';
 import '../templates/dart_dart_mappable_dto_template.dart';
 import '../templates/dart_enum_dto_template.dart';
@@ -16,8 +18,6 @@ import '../templates/kotlin_retrofit_client_template.dart';
 import '../templates/kotlin_typedef_template.dart';
 import 'generated_file.dart';
 import 'json_serializer.dart';
-import 'universal_data_class.dart';
-import 'universal_rest_client.dart';
 
 /// Enumerates supported programming languages to determine templates
 enum ProgrammingLanguage {
