+++ swagger_parser/lib/src/generator/templates/dart_dart_mappable_dto_template.dart
@@ -1,12 +1,12 @@
 import 'package:collection/collection.dart';
 
+import '../../parser/model/universal_data_class.dart';
+import '../../parser/model/universal_type.dart';
 import '../../utils/case_utils.dart';
 import '../../utils/type_utils.dart';
 import '../../utils/utils.dart';
 import '../models/json_serializer.dart';
 import '../models/programming_language.dart';
-import '../models/universal_data_class.dart';
-import '../models/universal_type.dart';
 import 'dart_import_dto_template.dart';
 
 String dartDartMappableDtoTemplate(
