+++ swagger_parser/lib/src/generator/templates/kotlin_enum_dto_template.dart
@@ -1,8 +1,8 @@
 import 'package:collection/collection.dart';
 
+import '../../parser/model/universal_data_class.dart';
 import '../../utils/case_utils.dart';
 import '../../utils/utils.dart';
-import '../models/universal_data_class.dart';
 
 /// Provides template for generating kotlin enum DTO
 String kotlinEnumDtoTemplate(
