+++ swagger_parser/lib/src/generator/templates/kotlin_retrofit_client_template.dart
@@ -1,11 +1,11 @@
+import '../../parser/model/universal_request.dart';
+import '../../parser/model/universal_request_type.dart';
+import '../../parser/model/universal_rest_client.dart';
+import '../../parser/model/universal_type.dart';
 import '../../utils/case_utils.dart';
 import '../../utils/type_utils.dart';
 import '../../utils/utils.dart';
 import '../models/programming_language.dart';
-import '../models/universal_request.dart';
-import '../models/universal_request_type.dart';
-import '../models/universal_rest_client.dart';
-import '../models/universal_type.dart';
 
 /// Return file contents for kotlin retrofit client
 /// File contents are generated using universal rest client
