+++ swagger_parser/lib/src/parser/model/universal_component_class.dart
