+++ swagger_parser/lib/src/parser/model/universal_data_class.dart
