+++ swagger_parser/lib/src/parser/model/universal_enum_class.dart
