+++ swagger_parser/lib/src/parser/model/universal_request_type.dart
