+++ swagger_parser/lib/src/parser/model/universal_rest_client.dart
