+++ swagger_parser/lib/src/parser/swagger_parser_core.dart
@@ -4,4 +4,9 @@ library swagger_parser_core;
 export 'config/parser_config.dart';
 export 'exception/parser_exception.dart';
 export 'model/open_api_info.dart';
+export 'model/universal_data_class.dart';
+export 'model/universal_request.dart';
+export 'model/universal_request_type.dart';
+export 'model/universal_rest_client.dart';
+export 'model/universal_type.dart';
 export 'parser/open_api_parser.dart';
