+++ swagger_parser/lib/src/utils/type_utils.dart
@@ -1,4 +1,4 @@
-import '../generator/models/universal_data_class.dart';
+import '../parser/model/universal_data_class.dart';
 import 'case_utils.dart';
 import 'dart_keywords.dart';
