+++ swagger_parser/test/e2e/tests/enum_member_names/expected_files/export.dart
@@ -2,11 +2,9 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint, unused_import
 
-// Root client
-export 'rest_client.dart';
-
 // Clients
 export 'clients/api_client.dart';
-
 // Data classes
 export 'models/enum_class.dart';
+// Root client
+export 'rest_client.dart';
