+++ swagger_parser/test/e2e/tests/enum_member_names/expected_files/models/enum_class.dart
@@ -9,49 +9,38 @@ enum EnumClass {
   /// The name has been replaced because it contains a keyword. Original name: `-index`.
   @JsonValue('-index')
   valueMinusIndex('-index'),
-
   @JsonValue('-name')
   minusName('-name'),
 
   /// The name has been replaced because it contains a keyword. Original name: `index`.
   @JsonValue('index')
   valueIndex('index'),
-
   @JsonValue('name')
   name('name'),
 
   /// The name has been replaced because it contains a keyword. Original name: `json`.
   @JsonValue('json')
   valueJson('json'),
-
   @JsonValue('yaml')
   yaml('yaml'),
-
   @JsonValue('-1')
   valueMinus1('-1'),
-
   @JsonValue('0')
   value0('0'),
-
   @JsonValue('1')
   value1('1'),
-
   @JsonValue('1itemOne')
   value1itemOne('1itemOne'),
-
   @JsonValue('2ItemTwo')
   value2ItemTwo('2ItemTwo'),
-
   @JsonValue('3item_three')
   value3itemThree('3item_three'),
-
   @JsonValue('4ITEM-FOUR')
   value4ItemFour('4ITEM-FOUR'),
 
   /// Incorrect name has been replaced. Original name: `5иллегалчарактер`.
   @JsonValue('5иллегалчарактер')
   undefined0('5иллегалчарактер'),
-
   @JsonValue('6 item six')
   value6ItemSix('6 item six'),
