+++ swagger_parser/test/e2e/tests/multipart_request_properties/expected_files/export.dart
@@ -2,11 +2,9 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint, unused_import
 
-// Root client
-export 'rest_client.dart';
-
 // Clients
 export 'clients/client_client.dart';
-
 // Data classes
 export 'models/object0.dart';
+// Root client
+export 'rest_client.dart';
