+++ swagger_parser/test/generator/root_client_test.dart
@@ -1,5 +1,5 @@
 import 'package:swagger_parser/src/generator/fill_controller.dart';
-import 'package:swagger_parser/src/generator/models/universal_rest_client.dart';
+import 'package:swagger_parser/src/parser/swagger_parser_core.dart';
 import 'package:test/test.dart';
 
 void main() {
