+++ swagger_parser/test/parser/data_classes_test.dart
@@ -1,7 +1,5 @@
 import 'package:path/path.dart' as p;
-import 'package:swagger_parser/src/generator/models/universal_data_class.dart';
-import 'package:swagger_parser/src/generator/models/universal_type.dart';
-import 'package:swagger_parser/src/parser/parser.dart';
+import 'package:swagger_parser/src/parser/swagger_parser_core.dart';
 import 'package:swagger_parser/src/utils/file_utils.dart';
 import 'package:test/test.dart';
 
@@ -12,7 +10,7 @@ void main() {
           p.join('test', 'parser', 'schemas', 'empty_class.2.0.json');
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final actualDataClass = parser.parseDataClasses().first;
       const expectedDataClass = UniversalComponentClass(
         name: 'ClassName',
@@ -27,7 +25,7 @@ void main() {
           p.join('test', 'parser', 'schemas', 'empty_class.3.0.json');
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final actualDataClass = parser.parseDataClasses().first;
       const expectedDataClass = UniversalComponentClass(
         name: 'ClassName',
@@ -44,7 +42,7 @@ void main() {
           p.join('test', 'parser', 'schemas', 'basic_types_class.2.0.json');
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final actualDataClass = parser.parseDataClasses().first;
       const expectedDataClass = UniversalComponentClass(
         name: 'ClassName',
@@ -152,7 +150,7 @@ void main() {
           p.join('test', 'parser', 'schemas', 'basic_types_class.3.0.json');
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final actualDataClass = parser.parseDataClasses().first;
       const expectedDataClass = UniversalComponentClass(
         name: 'ClassName',
@@ -260,7 +258,7 @@ void main() {
           p.join('test', 'parser', 'schemas', 'reference_types_class.2.0.json');
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final dataClasses = parser.parseDataClasses().toList();
       const expectedDataClasses = [
         UniversalComponentClass(
@@ -311,7 +309,7 @@ void main() {
           p.join('test', 'parser', 'schemas', 'reference_types_class.3.0.json');
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final dataClasses = parser.parseDataClasses().toList();
       const expectedDataClasses = <UniversalDataClass>[
         UniversalComponentClass(
@@ -362,7 +360,7 @@ void main() {
           p.join('test', 'parser', 'schemas', 'of_like_class.3.1.json');
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final dataClasses = parser.parseDataClasses().toList();
       final expectedDataClasses = <UniversalDataClass>[
         const UniversalComponentClass(
@@ -460,7 +458,7 @@ void main() {
       );
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final dataClasses = parser.parseDataClasses().toList();
       final expectedDataClasses = <UniversalDataClass>[
         const UniversalComponentClass(
@@ -518,7 +516,7 @@ void main() {
       );
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final dataClasses = parser.parseDataClasses().toList();
       final expectedDataClasses = <UniversalDataClass>[
         const UniversalComponentClass(
@@ -568,7 +566,7 @@ void main() {
       final schemaPath = p.join('test', 'parser', 'schemas', 'enum_class.json');
       final configFile = schemaFile(schemaPath);
       final schemaContent = configFile!.readAsStringSync();
-      final parser = OpenApiParser(schemaContent);
+      final parser = OpenApiParser(ParserConfig(schemaContent, isJson: true));
       final dataClasses = parser.parseDataClasses();
       expect(dataClasses, hasLength(2));
       final enumClass = dataClasses.whereType<UniversalEnumClass>().first;
